+++ src/main/preload.ts
@@ -52,7 +52,6 @@ export type Channels =
   | 'reconfigureOverlay'
   | 'reconfigureCloud'
   | 'runDiskSizeMonitor'
-  | 'diskSizeMonitorComplete'
   | 'getSensibleEncoderDefault'
   | 'refreshCloudGuilds';